Run all Behat features in isolation
The initialization config for the tests are now using the same config
as middleware is using when a new environment is bootstrapped.

After the initial setup of the tests a mysqldump is created in
order to run all test in isolation and prevent unwanted side effects.

And the event stream is no only used to boostrap an SRAA with a token.