Make devconf installable again
Prior to this change the init script would not work because the migrations failed.
The migrations failed because the `doctrine:schema:create` command was used. This command creates the schema based on the current entity mapping, which is more or less the endpoint of the migrations.
So the migrations would fail. Ideally, `doctrine:migrations:migrate` would be used instead of `d:s:c`, but that is out of scope for this pr.